tests/wpt/tests/html/semantics/permission-element/usermedia/no-children-rendered.tentative.html

<!DOCTYPE html>
<meta charset=utf-8>
<link rel="help" href="https://github.com/WICG/PEPC/blob/main/explainer.md#locking-the-pepc-style">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<!-- The usermedia element is not a void element. Its children do not render. -->

<!-- This tests for implementations that used to have no end tag for the usermedia element
 In those implementations there would be 2 usermedia elements in this div. -->
<div id="usermedia-element-div">
  <usermedia type="camera">
  <usermedia type="microphone">
</div>

<usermedia id="mic" type="microphone">This is some text</usermedia>

<!-- The usermedia element does not automatically close <p> tags -->
<p id="paragraph">Foo <usermedia type="camera">bar</usermedia> baz</p>

<script>
  test(function(){
    assert_equals(1, document.getElementById("usermedia-element-div").childElementCount);
    assert_equals('', document.getElementById("mic").innerText);
    assert_equals('This is some text', document.getElementById("mic").textContent);
    assert_equals('This is some text', document.getElementById("mic").innerHTML);
    assert_equals(4, document.body.childElementCount); //div, usermedia, script, p
    assert_equals('Foo bar baz', document.getElementById("paragraph").textContent);
  }, "The usermedia element should have no end tag or contents");
</script>
</body>
